Fix managesf replication list/get-all subcommands
The commands list and get-all are basically broken in the REST API.

In the REST API it is not possible to get a list of configs with a URL like this:

http://managesf.tests.dom/manage/replication

There is always a sectionname and setting required. The following URL thus
works:

http://managesf.tests.dom/manage/replication/sectionname/setting

However, the CLI always calls a sys.exit(0) if the command "list" is used; and
even if this is fixed there is sometimes a "None" sent as section name and/or
setting. This is obviously not the right way to do it, so let's fix this and
make the REST API itself handle URLs without a section and setting name.

With this patch the following URLs are all valid and return expected data:

List of complete config:
http://managesf.tests.dom/manage/replication/

List only settings of section "sectionname":
http://managesf.tests.dom/manage/replication/sectionname
http://managesf.tests.dom/manage/replication/sectionname/

Only return setting in sectionname
http://managesf.tests.dom/manage/replication/sectionname/setting

Change-Id: Ia09ae112eda79d9aa38e51e78dc1bbf6083e4f71