Add sshconfig REST API
Add a basic interface to add or remove SSH configurations. Commands have to be
sent to the endpoint /sshconfig/<aliasname>/. A JSON dictionary in the following
format has to be sent with a PUT request to create or update a hostentry:

{
    "hostname": "Hostname",
    "identityfile_content": "TheActualKeyOfTheHost",
    "userknownhostsfile": "UserKnownHostsFile",
    "preferredauthentications": "PreferredAuthentications",
    "stricthostkeychecking": "StrictHostKeyChecking",
    "username": "Username"
}

The identityfile_content is written to a keyfile, and the keyfile name is added
using the setting "identityfile" to the host configuration.

The following is an example how to use curl to add a new ssh configuration named
"sample", using the example JSON from above saved in a file named "sample.json":

    curl -v -b ./cookiefile -c ./cookiefile -X POST
        --data "username=user1&password=userpass&back=/"
        http://tests.dom/auth/login

    curl -v -b ./cookiefile -c ./cookiefile -X PUT \
        http://tests.dom/manage/sshconfig/sample \
        -H 'Content-Type:application/json' \
        --data-binary "@sample.json"

The managesf CLI has been updated too and includes two commands to add or delete
a SSH config entry. Example usage:

    ssh-keygen -f private_ssh.key

    sfmanager --url http://managesf.tests.dom \
        --auth http://managesf.tests.dom \
        --auth user1:userpass \
        gerrit_ssh_config add --alias aliasname --key private_ssh.key --hostname sample.github.com

    sfmanager --url http://managesf.tests.dom \
        --auth http://managesf.tests.dom \
        --auth user1:userpass \
        gerrit_ssh_config delete --alias sample

Depends-On: Icdff122c034cb028376556e8d14ac8be60b7fe2a
Change-Id: Ic8bb85fc60d681fc11567b105f922eca50db737f