Add sshconfig REST API
Add a basic interface to add or remove SSH configurations. Commands have to be
sent to the endpoint /sshconfig/<aliasname>/. A JSON dictionary in the following
format has to be sent with a PUT request to create or update a hostentry:

{
    "hostname": "Hostname",
    "identityfile_content": "TheActualKeyOfTheHost",
    "userknownhostsfile": "UserKnownHostsFile",
    "preferredauthentications": "PreferredAuthentications",
    "stricthostkeychecking": "StrictHostKeyChecking",
    "username": "Username"
}

The identityfile_content is written to a keyfile, and the keyfile name is added
using the setting "identityfile" to the host configuration.

Change-Id: Ic8bb85fc60d681fc11567b105f922eca50db737f