Service plugin: redmine (WIP)
First step to introduce the new service plugin architecture. In this patch
we add a plugin for Redmine as served through Software Factory, ie with
cauth authentication - please note that backup/restore operations are not
currently covered by the plugin as they should.
The plugin is not dynamically loaded at the moment, but replaces the
old codebase managing redmine.

Change-Id: I7c70cbe7efb066dc262c32d8366768a5775020ac