backup in service plugins
Move the backup logic in each service. It is assumed that a service's
puppet deployment recipe defines two scripts called 'backup_%service%.sh'
and 'restore_%service%.sh' that can be called to take care of backups
and restoration operations. Additionnally, a heartbeat command can
be defined to ensure the service is running correctly after each operation.

Depends-On: Id181ce72d2e56085de9c918a9145949186726779

Change-Id: I6a24583a43de58d6d71534ebcfad36c7fc8c3cbb