Add a user backend in manageSF
Instead of relying on redmine for most user-related operations, we
now use a specific user database. This commit introduces the first
table related to users.
This table also tracks a user's external id as passed from the SSO
(cauth in our case) so that users can be tracked accurately despite
possible changes.

Related-To: #1235
Change-Id: I45cc36fb63ed9ad82119d711faa1f0252771c225
Depends-On: I520e8bb08062c5bab45525224c6b6dd9a91133c8