Add a mapping table between SF users and service users
By linking a SF user explicitly and uniquely to its corresponding
user in each service, it will be possible later on to update users
from a single REST endpoint, or to handle changes in user info
and propagate them back to the services.

As a bonus, use a context manager to handle sessions and avoid
running out of connection queues to mariadb.

Change-Id: I6558956b4ea999e503a21178487880d88546245a
Related-To: #1235
Depends-On: I7666eea3f3815dfd2d703f84c29700adc5ea034d