Remove replication and sshconfig REST endpoint
This commit is part of the replication refactoring.
The underlying idea is to let user propose change
in replication.config via the config repo.

See: http://softwarefactory-project.io/r/#/c/3540/

And in case of changes needed in gerrit known hosts file as
well as in .ssh/config an admin of SF should take care
of them then restart Gerrit manually.

Change-Id: I639de7a483ba9b37de881c441de7d5ecb018db3f