Allow membership endpoint usage to add groups to project groups
This patch keeps the actual API of managesf but accept a
group name as well as a user email to identify a group or
a user. An internal evaluation is done to check if we need to
deal with a user or a group.

This patch is then still compatible with sfmanager.
The user of sfmanager can use the membership command along
with the --user to identify both a user or a group.

Change-Id: Ibe553023ab28f48948841cd5c45ee859ac7614a0
Depends-on: If9ae7887cf110017993c9dc01909c7eebfb057ed
Depends-on: I7cc0022572c27179a3bf91d0efb15b2b2b71d197