Allow membership endpoint usage to add groups to project groups
- Standalone groups can be added/removed
- Project groups can be added/removed

This patch keeps the actual API of managesf but accept a
group name as well as a user email to identify a group or
a user. An internal evaluation is done to check if we need to
deal with a user or a group.

This patch is then still compatible with sfmanager.
The user of sfmanager can use the membership command along
with the --user to identify both a user or a group.

One important note about that patch:

The actual project group implementation on Redmine
prevent this patch to easily allow the addition and
removal of a project group in a project group in Redmine.
Indeed we don't have project group on redmine. So the
API will accept this kind of request, handle it for
Gerrit but will skip Redmine.

Change-Id: Ibe553023ab28f48948841cd5c45ee859ac7614a0
Depends-on: If9ae7887cf110017993c9dc01909c7eebfb057ed
Depends-on: I7cc0022572c27179a3bf91d0efb15b2b2b71d197