Add XXXImpact redmine hook after merge
Adding "XXXImpact" (where XXX belongs to [a-zA-Z0-9_]) on a single
line in the commit message will trigger this hook and create an
issue on Redmine to remind that more work is needed. One issue
is created per (unique) occurrence of XXXImpact.

Example: when a patch with the following commit message is merged:

"Add REST API for ordering ramen

DocImpact
GUIImpact"

the following issue will then be created:

* [Doc] Add REST API for ordering ramen
* [GUI] Add REST API for ordering ramen

with description linking to the patch's review page.

Change-Id: I0ee16dad36e4c02c53c8887b04ee09d4c727b26a
Related-To: #1446
Depends-On: Ib61ccf600226278a71333fab7983f20f89688153