Add storyboard service
This change adds a new service to manage storyboard
user. Because the storyboard webclient requires
an user_id, this service will force the storyboard userid
to be identical to cauth userid. This allows the webclient
to use the cauth userid provided by auth_pub_tkt cookie.

Moreover, because the storyboard api is already protected
by cauth, a fake token set to username is used so that the
webclient believe it is authenticated.

Issue: #1455
Change-Id: I71240f00ca6b9d91050395ebcaea10064aab879c