Managesf: resources - add a direct apply mode for PUT
PUT can be called to apply resources from a YAML str passed
as json body of the PUT request.

Resources YAML structure/content is validated and then apply
on services.

This call is not part of the classic workflow and won't be
documented. Its main usage is for bootstrapping resources
via managesf like the config project.

Change-Id: I61c104b33cdfa78a7ff8ee9880a17255731d768b