Resource GIT style : Provide endpoint to get missing resources
* Add gerrit operations for reading:
- gitrepos
- acls
- groups

* Add engine method to return a resources struct with only
resources that exists on the services (gerrit) but not
references in the config/resources/ tree.

The return struct will be use to open a new review
aginst the config repo that will re-sync both the
config repo and the reality. This new review will contains
a tag in its commit message to tell the config-update to
skip running the callbacks on services.

* Some special groups are skipped by the Group resources:
- Administrators
- Non Interactive Users

The Administrator group should be managed by a specific group
resources IMO. So it will be taken in account in a follow up
patch.

Change-Id: I003acea6459e661f5ae40925a9b8f64b51d0d00b