Agents (Nodepool) REST API
This commit introduces a REST API endpoint at /agents/ that covers
operations related to slaves created on the fly by an agent manager,
namely nodepool in the context of Software Factory.

The REST API uses a service plugin for Nodepool. Due to Nodepool's
lack of a REST API at the moment, the plugin executes remote calls
to Nodepool's CLI on the relevant SF node via SSH.

Change-Id: I847185f5514f6c0703d382d3840a1e551f2f3e36