resources: modify the apply checkout strategy to run no matter the tips of HEAD
This patch changes the strategy where we expect to have HEAD^1
the previous state of the resources tree and HEAD the wanted state.

In a case of something has been merged on the config repo
not related to the resources then the apply will not
able to detect any changes because HEAD^1 will be wanted state
and HEAD will be also the wanted state.

So this patch only search for changes on the resources/ tree and
will checkout last_change_on_resources^1. This means a manual
call to /usr/local/bin/resources.sh apply can be done whatever
the tip of config and will try to apply the changes between
HEAD and last_change_on_resources^1.

Note: Not sure this a really valuable change ... need to
discuss on that.

Change-Id: Ic6c498aaddfc9b33a314b7cf97079d57b286a3c6